Move unquoteString closer to LambdaParser (#266)
Because it's used only there
