+++ kernel/src/main/scala/me/rexim/morganey/syntax/Language.scala
@@ -1,5 +1,9 @@
 package me.rexim.morganey.syntax
 
+import hiddenargs.{hidden, hiddenargs}
+
+import scala.annotation.tailrec
+
 object Language {
 
   val identifier         = "[a-zA-Z][a-zA-Z0-9]*"
@@ -49,4 +53,21 @@ object Language {
 
   val rangeOperator      = ".."
 
+  @hiddenargs
+  @tailrec
+  def unquoteString(s: String, @hidden acc: String = ""): String =
+    if (s.isEmpty) acc else s(0) match {
+      case '"' => unquoteString(s.tail, acc)
+      case '\\' => s(1) match {
+        case 'b' => unquoteString(s.drop(2), acc + "\b")
+        case 'f' => unquoteString(s.drop(2), acc + "\f")
+        case 'n' => unquoteString(s.drop(2), acc + "\n")
+        case 'r' => unquoteString(s.drop(2), acc + "\r")
+        case 't' => unquoteString(s.drop(2), acc + "\t")
+        case '"' => unquoteString(s.drop(2), acc + "\"")
+        case 'u' => unquoteString(s.drop(6), acc + Integer.parseInt(s.slice(2, 6), 16).toChar.toString)
+        case c   => unquoteString(s.drop(2), acc + c.toString)
+      }
+      case c => unquoteString(s.tail, acc + c.toString)
+    }
 }
